% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{plays}
\alias{plays}
\title{Extract the play-by-play data component from a volleyxml object}
\usage{
plays(x)
}
\arguments{
\item{x}{volleyxml: a volleyxml object as returned by \code{vx_read}}
}
\value{
The plays component of x (a data.frame)
}
\description{
Extract the play-by-play data component from a volleyxml object
}
\examples{
\dontrun{
  x <- vx_read(vx_example_file())
  head(plays(x))
}
}
\seealso{
\code{\link{vx_read}}
}
