% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{summary.volleyxml}
\alias{summary.volleyxml}
\title{A simple summary of a volleyball match}
\usage{
\method{summary}{volleyxml}(object, ...)
}
\arguments{
\item{object}{volleyxml: volleyxml object as returned by \code{vx_read}}

\item{...}{: additional arguments (currently these have no effect)}
}
\value{
list of summary items
}
\description{
A simple summary of a volleyball match
}
\examples{
x <- vx_read(vx_example_file())
summary(x)

}
\seealso{
\code{\link{vx_read}}
}
