% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{teams}
\alias{teams}
\alias{home_team}
\alias{visiting_team}
\title{Get team names from volleyxml object}
\usage{
teams(x)

home_team(x)

visiting_team(x)
}
\arguments{
\item{x}{volleyxml or data.frame: a volleyxml object as returned by \code{vx_read}, or the plays component of that object}
}
\value{
character vector of team names
}
\description{
Get team names from volleyxml object
}
\examples{
\dontrun{
  x <- vx_read(vx_example_file())
  teams(x)
}
}
\seealso{
\code{\link{vx_read}}
}
