% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volleyxml.R
\docType{package}
\name{volleyxml}
\alias{volleyxml}
\alias{volleyxml-package}
\title{\pkg{volleyxml}}
\description{
Provides basic functions for parsing volleyball scouting files in XML format.
}
\author{
Ben Raymond \email{ben@untan.gl}
}
