% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{vx_example_file}
\alias{vx_example_file}
\title{Example data files}
\usage{
vx_example_file(choice = 1)
}
\arguments{
\item{choice}{numeric: which data file to return?
\itemize{
  \item{1 - a largely fictitious example}
}}
}
\value{
path to the file
}
\description{
Example data files provided as part of the volleyxml package
}
\examples{
myfile <- vx_example_file()
x <- vx_read(myfile)
summary(x)

}
\seealso{
\code{\link{vx_read}}
}
