% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{vx_read}
\alias{vx_read}
\title{Read XML volleyball scouting file}
\usage{
vx_read(filename)
}
\arguments{
\item{filename}{string: name of the file to read}
}
\value{
volleyxml object (list with components meta, plays, and raw)
}
\description{
Read XML volleyball scouting file
}
\examples{
x <- vx_read(vx_example_file())

}
